HW 7 update 1
I relinked the proper css file (garden2)
I have changed the fonts to white, but I can’t seem to get the image
larger? I have it within a container (main) but it is not responding to
css commands I thought might enlargen it. I am done for the night. Need
sleep.