4:02pm Tuesday HW 7 DONE!
I used padding and margin to adjust the size of the anchor “border” or
“container” if you will, and adjusted the color the closest I could to
what was shown. I didn’t see a yellow-green that matched the
assignment, so I chose something close that allowed for clarity and
visibility.  I had the wrong image for the BG so I adjusted that, after
trying to use linear-gradients to darken the large version of image 1.
Whoops. Done!