10:58am HW7 Wed. DONE!!
This is about as good as I know how to get it.

1. Created a DIV container for both anchors, so I could use margin left
to align them with the above text.
Removed text align center to allow this in CSS for this div class.

2. Did the same with the above text and their respective text/s. Margin
left 25px did the trick. Then I simply reduced padding top to draw them
closer together on the anchor div.

3. Reduced font size to look a bit better on the left divs, with the
longer coloring, from 20 to 15. Looks better I think.

4. Celebrate.