3/1/16 7:02pm Tuesday HW8
I finished it! Thank GOD.

Okay, so I redid the whole bottom section, with a different method.
Rather than setting H&W for the section, I used the section as a ID in
case I needed it, but I created 3 DIV classes for each wine bottle, and
their respective h1’s and p’s. I simply set 33% width for each, and
adjusted CSS styling for the anchors and text, while also adjusting the
margin and padding.

Lastly, I created a DIV class for the hero splash h1 and p’s, so I
could center align the DIV and more easily move it up and down using
margin and padding.

This isn’t exactly like the homework, in terms of visual matching, but
the heart and soul is there. I used a little of my own creativity on
it, and the placement is there. It’s certainly miles beyond the mess I
had yesterday. It was important for me to show I could do this one
without any help from home.