HW10. 3/24/16. 8:58pm
I am stuck here. I don’t know what the problem is, and I feel like I
have tried EVERYTHING. My aside is stuck outside of the BODY section,
and as a result can’t move up. In the HTML in normal doc flow, the BODY
contains the section, aside and div, so there is no reason for the
yellow section to be outside of it, nor is there any reason for me to
not be able to move it up. I can’t move forward on this assignment
until I fix this, and I can’t figure it out. I have reached out on
slack, but help is currently not available due to instructing. I am
going to resume in the morning after someone is able to look at my code.