HW1. 3/29/16 6:22pm
I went back and polished the original day 1 HW as it was quite awful. I
didn’t want to get too fancy, as the final look should I feel be
something that we would be capable or close to capable of in the first
week, so nothing too fancy, no bootstrap, etc. Just a simple, cleaner
look and design, but again, basic, as it was week 1.