HW2. 3/29/16 8:37pm
I have re-done the entire page now, in light of what would be skill
appropriate for those early weeks, with an eye not to get too fancy.

The original site, while the 2nd thing I ever did, was a real eyesore
to me, so I did a simple, clean redo of it according to early course
skill levels.